+++ lib/flows/plugin/profiler/wrapper.rb
@@ -4,21 +4,45 @@ module Flows
       # @api private
       module Wrapper
         class << self
-          def make_module(klass, method_type, method_name) # rubocop:disable Metrics/MethodLength
+          def make_instance_wrapper(method_name) # rubocop:disable Metrics/MethodLength
             Module.new.tap do |mod|
-              mod.define_method(method_name) do |*args, &block|
+              mod.define_method(method_name) do |*args, &block| # rubocop:disable Metrics/MethodLength
+                thread = Thread.current
+                klass = self.class
+
+                return super(*args, &block) unless thread[THREAD_VAR_FLAG]
+
+                report = thread[THREAD_VAR_REPORT]
+                report.add(:started, klass, :instance, method_name, nil)
+
+                before = Process.clock_gettime(Process::CLOCK_MONOTONIC, :float_microsecond)
+                super(*args, &block)
+              ensure
+                if before
+                  after = Process.clock_gettime(Process::CLOCK_MONOTONIC, :float_microsecond)
+                  report.add(:finished, klass, :instance, method_name, after - before)
+                end
+              end
+            end
+          end
+
+          def make_singleton_wrapper(method_name) # rubocop:disable Metrics/MethodLength
+            Module.new.tap do |mod|
+              mod.define_method(method_name) do |*args, &block| # rubocop:disable Metrics/MethodLength
                 thread = Thread.current
 
                 return super(*args, &block) unless thread[THREAD_VAR_FLAG]
 
                 report = thread[THREAD_VAR_REPORT]
-                report.add(:started, klass, method_type, method_name, nil)
+                report.add(:started, self, :singleton, method_name, nil)
 
                 before = Process.clock_gettime(Process::CLOCK_MONOTONIC, :float_microsecond)
                 super(*args, &block)
               ensure
-                after = Process.clock_gettime(Process::CLOCK_MONOTONIC, :float_microsecond)
-                report.add(:finished, klass, method_type, method_name, after - before)
+                if before
+                  after = Process.clock_gettime(Process::CLOCK_MONOTONIC, :float_microsecond)
+                  report.add(:finished, self, :singleton, method_name, after - before)
+                end
               end
             end
           end