+++ spec/flows/plugin/profiler/report/raw_spec.rb
@@ -1,16 +1,16 @@
 require 'spec_helper'
 
 RSpec.describe Flows::Plugin::Profiler::Report::Raw do
-  subject(:raw_report) { described_class.new }
+  subject(:report) { described_class.new }
 
   let(:klass) { Class.new }
 
   describe '#add' do
-    subject(:raw_data) { raw_report.raw_data }
+    subject(:raw_data) { report.raw_data }
 
     before do
-      raw_report.add(:started, klass, :singleton, :perform, nil)
-      raw_report.add(:finished, klass, :singleton, :perform, 10.5)
+      report.add(:started, klass, :singleton, :perform, nil)
+      report.add(:finished, klass, :singleton, :perform, 10.5)
     end
 
     it do
@@ -22,13 +22,29 @@ RSpec.describe Flows::Plugin::Profiler::Report::Raw do
   end
 
   describe '#to_s' do
-    subject(:text) { raw_report.to_s }
+    subject(:text) { report.to_s }
 
     before do
-      raw_report.add(:started,  klass, :singleton, :perform, nil)
-      raw_report.add(:finished, klass, :singleton, :perform, 10.5)
+      report.add(:started,  klass, :singleton, :perform, nil)
+      report.add(:finished, klass, :singleton, :perform, 10.5)
     end
 
     it { is_expected.to be_a String }
+  end
+
+  describe '#events' do
+    subject(:events) { report.events }
+
+    before do
+      report.add(:started,  klass, :singleton, :perform, nil)
+      report.add(:finished, klass, :singleton, :perform, 10.5)
+    end
+
+    it 'returns array of events' do
+      expect(events).to match([
+                                instance_of(described_class::StartEvent),
+                                instance_of(described_class::FinishEvent)
+                              ])
+    end
   end
 end